requirements.txt
python-telegram-bot
python-dotenv
requests
beautifulsoup4
lxml
apify-client
